LSTMStockTrading/requirements.txt
# Production Requirements
# LSTM Stock Trading System v2.0

# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data & API
requests>=2.31.0
python-dotenv>=1.0.0
yfinance>=0.2.28

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0  # Advanced configuration management (optional)

# Experiment tracking
mlflow>=2.7.0
tensorboard>=2.13.0

# Plotting & visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Performance & utilities
joblib>=1.3.0
tqdm>=4.65.0
numba>=0.57.0  # Just-in-time compilation for speedups

# Statistical analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Data formats
pyarrow>=12.0.0  # For Parquet files
fastparquet>=2023.4.0

# Testing (move to requirements-dev.txt for production)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# pytest-mock>=3.11.0

# Code quality (move to requirements-dev.txt for production)
# black>=23.7.0
# isort>=5.12.0
# pylint>=2.17.0
# mypy>=1.4.0
